+++ Source/BackStreet/Character/public/CharacterBase.h
@@ -59,7 +59,7 @@ public:
 
 	//디버프 데미지를 입힘 (일회성)
 	UFUNCTION(BlueprintCallable)
-		float TakeDebuffDamage(float DamageAmount, uint8 DebuffType, AActor* Causer);
+		float TakeDebuffDamage(float DamageAmount, ECharacterDebuffType DebuffType, AActor* Causer);
 
 	//플레이어가 체력을 회복함 (일회성)
 	UFUNCTION()
@@ -72,15 +72,11 @@ public:
 		void InitCharacterState();
 
 	//캐릭터의 버프/디버프 정보를 업데이트
-	virtual	bool AddNewBuffDebuff(bool bIsDebuff, uint8 BuffDebuffType, AActor* Causer = nullptr, float TotalTime = 0.0f, float Value = 0.0f);
+	virtual	bool AddNewDebuff(ECharacterDebuffType BuffDebuffType, AActor* Causer = nullptr, float TotalTime = 0.0f, float Value = 0.0f);
 
 	//디버프가 활성화 되어있는지 반환
 	UFUNCTION(BlueprintCallable, BlueprintPure)
 		bool GetDebuffIsActive(ECharacterDebuffType DebuffType);
-
-	//버프가 활성화 되어있는지 반환
-	UFUNCTION(BlueprintCallable, BlueprintPure)
-		bool GetBuffIsActive(ECharacterBuffType BuffType);
 
 	//캐릭터의 스탯을 업데이트
 	UFUNCTION(BlueprintCallable)
