+++ Source/BackStreet/Global/private/BackStreetGameModeBase.cpp
@@ -1,7 +1,7 @@
 // Copyright Epic Games, Inc. All Rights Reserved.
 
 #include "../public/BackStreetGameModeBase.h"
-#include "../public/BuffDebuffManager.h"
+#include "../public/DebuffManager.h"
 #include "../../StageSystem/public/GridBase.h"
 #include "../../StageSystem/public/TileBase.h"
 #include "../../StageSystem/public/ChapterManagerBase.h"
@@ -37,9 +37,8 @@ void ABackStreetGameModeBase::BeginPlay()
 	PlayerCharacterRef = Cast<AMainCharacterBase>(UGameplayStatics::GetPlayerCharacter(GetWorld(), 0));
 
 	//------ Global Buff/Debuff Manager √ ±‚»≠ --------
-	BuffDebuffManager = NewObject<UBuffDebuffManager>(this, UBuffDebuffManager::StaticClass(), FName("BuffDebuffManager"));
-	UE_LOG(LogTemp, Warning, TEXT("TryActivate BuffManager #1"));
-	BuffDebuffManager->InitBuffManager(this);
+	DebuffManager = NewObject<UDebuffManager>(this, UDebuffManager::StaticClass(), FName("BuffDebuffManager"));
+	DebuffManager->InitDebuffManager(this);
 }
